Fixed code repetition in Button and FileCard (component is now called DropdownButton) #93
Also includes a lot of renaming of files and seperated Grid and List components into different folders.
