+++ RC-SpeechToText/ClientApp/components/Dashboard/DropdownButton.tsx
@@ -1,20 +1,18 @@
-import * as React from 'react';
+﻿import * as React from 'react';
 import axios from 'axios';
 import auth from '../../Utils/auth';
-import { Link } from 'react-router-dom';
 import { ModifyTitleModal } from '../Modals/ModifyTitleModal';
 import { ModifyDescriptionModal } from '../Modals/ModifyDescriptionModal';
 import { SuccessModal } from '../Modals/SuccessModal';
 import { ErrorModal } from '../Modals/ErrorModal';
-import { LoadingModal } from '../LoadingModal';
 
 interface State {
+    fileId: number,
     title: string,
     description: string,
     modifiedTitle: string,
     newDescription: string,
     flag: string,
-    loading: boolean,
     showDropdown: boolean,
     showTitleModal: boolean,
     showDescriptionModal: boolean,
@@ -26,17 +24,17 @@ interface State {
     unauthorized: boolean
 }
 
-export class FileCard extends React.Component<any, State> {
+export class DropdownButton extends React.Component<any, State> {
     constructor(props: any) {
         super(props);
 
         this.state = {
+            fileId: this.props.fileId,
             title: this.props.title,
             description: this.props.description,
             modifiedTitle: "",
             newDescription: "",
             flag: this.props.flag,
-            loading: false, 
             showDropdown: false,
             showTitleModal: false,
             showDescriptionModal: false,
@@ -58,7 +56,6 @@ export class FileCard extends React.Component<any, State> {
 
     // Add event listener for a click anywhere in the page
     componentDidMount() {
-        this.setState({ description: this.state.description }); 
         document.addEventListener('mouseup', this.hideDropdown);
     }
     // Remove event listener
@@ -69,7 +66,6 @@ export class FileCard extends React.Component<any, State> {
     public deleteWords = () => {
 
         console.log("Deleting words");
-        this.setState({ 'loading': true });
 
         const config = {
             headers: {
@@ -77,20 +73,14 @@ export class FileCard extends React.Component<any, State> {
                 'content-type': 'application/json'
             }
         }
-        axios.delete('/api/word/DeleteWordsByFileId/' + this.props.fileId, config)
+        axios.delete('/api/word/DeleteByFileId/' + this.state.fileId, config)
             .then(res => {
                 console.log(res.status);
                 this.deleteVersion();
             })
             .catch(err => {
                 if (err.response.status == 401) {
-                    this.setState({ 'loading': false });
-                    this.showErrorModal("Supprimer le ficher", "Veuillez vous connecter avant de supprimer le fichier.");
                     this.setState({ 'unauthorized': true });
-                }
-                else if (err.response.status == 400) {
-                    this.setState({ 'loading': false });
-                    this.showErrorModal("Supprimer le ficher", err.response.data);
                 }
             });
     }
@@ -106,20 +96,14 @@ export class FileCard extends React.Component<any, State> {
             }
         }
 
-        axios.delete('/api/version/DeleteFileVersions/' + this.props.fileId, config)
+        axios.delete('/api/version/DeleteFileVersions/' + this.state.fileId, config)
             .then(res => {
-                console.log(res.status);
+                console.log(res.data);
                 this.deleteFile();
             })
             .catch(err => {
                 if (err.response.status == 401) {
-                    this.setState({ 'loading': false });
-                    this.showErrorModal("Supprimer le ficher", "Veuillez vous connecter avant de supprimer le fichier.");
                     this.setState({ 'unauthorized': true });
-                }
-                else if (err.response.status == 400) {
-                    this.setState({ 'loading': false });
-                    this.showErrorModal("Supprimer le ficher", err.response.data);
                 }
             });
     };
@@ -131,22 +115,14 @@ export class FileCard extends React.Component<any, State> {
                 'content-type': 'application/json'
             }
         }
-        axios.delete('/api/file/Delete/' + this.props.fileId, config)
+        axios.delete('/api/file/Delete/' + this.state.fileId, config)
             .then(res => {
                 console.log(res.data);
-                this.setState({ 'loading': false });
-                this.showSuccessModal("Supprimer le ficher", "Le fichier intitulé '" + this.props.title + "' a été supprimé avec succès!");
-                this.props.updateFiles();
+                alert("File deleted");
             })
             .catch(err => {
                 if (err.response.status == 401) {
-                    this.setState({ 'loading': false });
-                    this.showErrorModal("Supprimer le ficher", "Veuillez vous connecter avant de supprimer le fichier.");
                     this.setState({ 'unauthorized': true });
-                }
-                else if (err.response.status == 400) {
-                    this.setState({ 'loading': false });
-                    this.showErrorModal("Supprimer le ficher", err.response.data);
                 }
             });
     }
@@ -167,11 +143,11 @@ export class FileCard extends React.Component<any, State> {
                 }
             }
 
-            axios.put('/api/file/ModifyTitle/' + this.props.fileId, formData, config)
+            axios.put('/api/file/ModifyTitle/' + this.state.fileId, formData, config)
                 .then(res => {
                     this.setState({ title: this.state.modifiedTitle });
                     this.hideTitleModal();
-                    this.showSuccessModal("Modifier le titre" , "Enregistrement du titre confirmé! Les changements effectués ont été enregistré avec succès.");
+                    this.showSuccessModal("Modifier le titre", "Enregistrement du titre confirmé! Les changements effectués ont été enregistré avec succès.");
                 })
                 .catch(err => {
                     console.log(err);
@@ -179,15 +155,15 @@ export class FileCard extends React.Component<any, State> {
                         this.showErrorModal("Modifier le titre", "Veuillez vous connecter avant de modifier le titre.");
                         this.setState({ 'unauthorized': true });
                     }
-                    else if (err.response.status == 500) {
-                        this.showErrorModal("Modifier le titre", err.response.data.message);
+                    else if (err.response.status == 400) {
+                        this.showErrorModal("Modifier le titre", err.response.data);
                     }
                 });
         }
         else {
             this.showErrorModal("Modifier le titre", "Enregistrement du titre annulé! Vous n'avez effectué aucun changements ou vous avez apporté les mêmes modifications.");
         }
-        
+
 
     }
 
@@ -196,7 +172,7 @@ export class FileCard extends React.Component<any, State> {
         var oldDescription = this.state.description
         var newDescription = this.state.newDescription
 
-        var modalTitle = (this.state.description && this.state.description != "" ? "Modifier la description" : "Ajouter une description") 
+        var modalTitle = (this.state.description && this.state.description != "" ? "Modifier la description" : "Ajouter une description")
 
         const formData = new FormData();
         formData.append("newDescription", newDescription)
@@ -209,7 +185,7 @@ export class FileCard extends React.Component<any, State> {
                 }
             }
 
-            axios.put('/api/file/saveDescription/' + this.props.fileId, formData, config)
+            axios.put('/api/file/saveDescription/' + this.state.fileId, formData, config)
                 .then(res => {
                     this.setState({ description: this.state.newDescription });
                     this.hideDescriptionModal();
@@ -274,7 +250,7 @@ export class FileCard extends React.Component<any, State> {
         this.setState({ modalTitle: title });
         this.setState({ showErrorModal: true });
     }
-    
+
     public hideSuccessModal = () => {
         this.setState({ showSuccessModal: false });
     }
@@ -285,51 +261,26 @@ export class FileCard extends React.Component<any, State> {
 
     public render() {
         return (
-            <div className="column is-3">
-                <div className="card fileCard">
-                    <span className={`tag is-rounded ${this.state.flag.indexOf("A") == 0 ? "is-danger" : this.state.flag.indexOf("R") == 0 ? "is-success" : "is-info"}`}>{this.state.flag}</span> 
-                    <header className="card-header">
-                        <p className="card-header-title fileTitle">
-                            {this.state.title ? (this.state.title.lastIndexOf('.') != -1 ? this.state.title.substring(0, this.state.title.lastIndexOf('.')) : this.state.title) : null}</p>
-                        <div className={`dropdown is-right ${this.state.showDropdown ? "is-active" : null}`} >
-                            <div className="dropdown-trigger">
-                                <div className="is-black" aria-haspopup="true" aria-controls="dropdown-menu4" onClick={this.showDropdown}>
-                                    <i className="fas fa-ellipsis-v "></i>
-                                </div>
-                            </div>
-                            <div className="dropdown-menu" id="dropdown-menu4" role="menu">
-                                <div className="dropdown-content">
-                                    <a className="dropdown-item" onClick={this.showTitleModal}>
-                                        Modifier le titre
-                                    </a>
-                                    <a className="dropdown-item" onClick={this.deleteWords}>
-                                        Effacer le fichier
-                                    </a>
-                                    {this.props.description ? <a className="dropdown-item" onClick={this.showDescriptionModal}>
-                                        Modifier la description
-                                    </a> : <a className="dropdown-item" onClick={this.showDescriptionModal}>
-                                            Ajouter une description
-                                    </a>}
-                                </div>
-                            </div>
-                        </div>
-                    </header>
-                    <div className="card-image">
-                        <div className="hovereffect">
-                            <figure className="image is-16by9">
-                                <img src={this.props.image} alt="Placeholder image" />
-                                <div className="overlay">
-                                    <Link className="info" to={`/FileView/${this.props.fileId}`}>View/Edit</Link>
-                                </div>
-                            </figure>
+            <div>
+                <div className={`dropdown is-right ${this.state.showDropdown ? "is-active" : null}`} >
+                    <div className="dropdown-trigger">
+                        <div className="is-black" aria-haspopup="true" aria-controls="dropdown-menu4" onClick={this.showDropdown}>
+                            <i className="fas fa-ellipsis-v "></i>
                         </div>
                     </div>
-                    <div className="card-content">
-                        <div className="content fileContent">
-                            <p className="transcription">{this.rawToWhiteSpace(this.props.transcription)}</p>
-                            <p><b>{this.props.username}</b></p>
-                            <time dateTime={this.props.date}>{this.props.date}</time>
-                            {/* <p><b>Description:</b> {this.state.description}</p> */}
+                    <div className="dropdown-menu" id="dropdown-menu4" role="menu">
+                        <div className="dropdown-content">
+                            <a className="dropdown-item" onClick={this.showTitleModal}>
+                                Modifier le titre
+                                    </a>
+                            <a className="dropdown-item" onClick={this.deleteWords}>
+                                Effacer le fichier
+                                    </a>
+                            {this.props.description ? <a className="dropdown-item" onClick={this.showDescriptionModal}>
+                                Modifier la description
+                                    </a> : <a className="dropdown-item" onClick={this.showDescriptionModal}>
+                                    Ajouter une description
+                                    </a>}
                         </div>
                     </div>
                 </div>
@@ -362,10 +313,6 @@ export class FileCard extends React.Component<any, State> {
                     hideModal={this.hideErrorModal}
                     title={this.state.modalTitle}
                     errorMessage={this.state.errorMessage}
-                />
-
-                <LoadingModal
-                    showModal={this.state.loading}
                 />
 
             </div>
